Update sample.py
Fixed RDF calculation: the results were giving the wrong order of magnitude and the units of the equation were not correct. 

Added the lacking "2dr" for 2D-Simulations.

Note: "len(system.distances)/(system.number_of_particles - 1)" is the same as "system.number_of_particles*0.5" but it was not clear to me where that came from.